refactor: prefer dedicated controls folder
